Bug 10278 - Add ability to hide items and records from search results for Independent Branches
For the staff intranet, enabling IndependentBranchesRecordsAndItems
will automatically add a branch limit filter to the search results
thus hiding all records without one or more items owned by the logged
in library. In add addition, all items whose homebranch is not that
libraries will be filtered and hidden from the search results and
record details.

This system preference will not affect the OPAC, unless the environment
variable BRANCHCODE is defined in the koha-httpd.conf file. If it is
defined, the same filters are applied to the OPAC, but based on the
branchcode value of the environment variable BRANCHCODE, rather than the
logged in branch.

Test Plan:
1) Apply patch
2) Run updatedatabase.pl
3) Perform a search that will give results for both records which
   have items owned by the logged in library, and records which have
   no items owned by the logged in library.
4) Enable the new system preference IndependentBranchesRecordesAndItems
5) Perform the same search again, any records without items owned
   by the currently logged in library should not appear.