Bug 10263 - Add ability to limit which branch can edit a bibliographic record ( IndependentBranchesMarcEditing )
Right now, it's possible to limit who can edit items by using
IndependantBranches. However, Koha has no method of limiting who can
edit bibliographic records, as they have no defined owner.

This patch adds a new system preference IndependentBranchesMarcEditing. If this new
syspref is enabled, a librarian can only edit or delete a bib record if
his or her logged in branch code matches the branchcode in
biblio.branchcode. This feature does not affect the ability to edit or
delete items, as that functionality is determined by
IndependantBranches.

This patch adds the 'branches_optional' field to the authorised values
pulldown for MARC fields and subfields in the MARC frameworks editor.
'branches_optional' behaves exactly like 'branches' except it includes
an "empty" choice so the branch does not have to be set.

Test Plan:
 1) Apply patch
 2) Run updatedatabase.pl
 3) Map a nonrepeatable field/subfield to biblio.branchcode ( e.g. 951$o )
    Pro-tip: Use the authorised value 'branches_optional' to get a pulldown of
    libraries instead of entering the branchcode manually.
 4) Edit a record, and add a branchcode.
 5) Log in as a non-superlibrarian with cataloging priveleges
 6) Change the logged in branch to a branch other than the one you
    entered in the record.
 7) You should note the 'edit' pulldown for the record is missing the
    "Edit record" and "Delete record" options.
 8) Switch the logged in branch to match the branchcode of the record.
 9) You should now see the edit and delete options for that record.
10) Attempt to edit a record with no branchcode defined, you should be
    able to.