Bug 11211 - Move calculation code out of C4::Calendar
This patch moves the calculation code out of C4::Calendar, replacing any
references to it with Koha::Calendar. Additionally, it moves the
administration interface to a single script, tools/calendar.pl.

NOTE: This is intended as a bridge to bug 8133, so the UI is a bit
awkward, and C4::Calendar has an API designed for adding hours to the
calendar. All features should still work, and 8133 will make things
better.

Test plan:
  1) Install the new dependency, Template::Plugin::JavaScript.
  2) prove t/db_dependent/Calendar.t t/db_dependent/Holidays.t to verify
     that the new C4::Calendar API works correctly.
  3) prove t/Calendar.t
  4) Add, remove and modify events of each type using the administrative
     interface. In this version, changing a repeating event to "Open"
     deletes it.
  5) Check out items to verify that the calculation code is using the
     events you just created.
  6) If possible, verify that the two affected cronjobs still function
     correctly.

Signed-off-by: wajasu <matted-34813@mypacks.net>