MDAH - Bug 8352 - Add printer administration, update script to use Koha defined printers
This patch adds the ability to define printers within Koha. The printer
CRUD is based on the code for the quote of the day system.

This patch extends the existing ability to print via LPR to any network
printer with the added ability to print to a CUPS printer configured on
the Koha server.

Test Plan:
1) Apply this patch set
2) Run updatedatabase.pl
3) Browse to admin/printers.pl
4) Configure a network printer via LPR ( explained in the printers help )
5) Create a hold
6) Execute the command "misc/cronjobs/holds/print_holds.pl -v --printer $BRANCHCODE=$PRINTER_ID
   where branchcode is the pickup branch for the hold, and printer_id is
   the id number for the printer.
7) See your reserve slip is printer!

Bonus points: Repeat the test plan, but for a printer you set up via CUPS on the Koha server