Bug 12725: Show RDA tag 264 on Intranet Results and Detail XSLT view (MARC21)
As bugs 12724 and 12726 did on OPAC, this patch adds similar changes to
Intranet Results and Details XSLT view. It shows information from tag 264 on
the Publisher line of Intranet Results and Detail.

Test plan (see also bug 12742 and the rules mentioned there):
[1] Add tag 260 and some 264s to your MARC21 record.
[2] Check the display of Intranet Results.
[3] Check the display of Intranet Detail.

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>
Results and detail show 264. I added 2 but only shows the second,
is that right?

[Reply from marcel:] Yes. Based on the rules described, it picks the most
'appropriate' 264.

Signed-off-by: Brendan Gallagher <brendan@bywatersolutions.com>