Bug 10883 - AllowHoldDateInFuture can let a patron jump to the top of the holds list
If a patron is allowed to add a future date for the hold using the
"Hold starts on date" field, that patron can put a date in that is
in the *past* which will cause them to move to the top of the holds
queue ( assuming they make the past date far enough back that it is
earlier than all the other hold dates ).

Test Plan:
1) Enable future holds
2) Note you can use dates in the past for "hold starts on date"
3) Note you can use dates in the past for "hold expires on date"
4) Apply this patch
5) Refresh the page
6) Note you can no longer use dates in the past for those fields

Signed-off-by: Christopher Brannon <cbrannon@cdalibrary.org>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Changed subject: OpacAllowHoldDateInFuture to AllowHoldDateInFuture
as this fix applies to staff.
Works as described, passes tests and QA script.