Bug 13315 [Follow-up] Add feedback for last item checked out to circulation.pl
This patch changes the position and styling of the information about the
last checkout so that it is grouped with the checkout form itself,
placing the message in a more visible position, especially for patrons
with a lot of information in the right-hand messages column which might
push the checkouts table down offscreen.

To test, apply the patch and clear your browser cache in order to
refresh the main CSS file.

Check out an item and confirm that a message is prominently displayed.

Signed-off-by: Jason Burds  <jburds@dubuque.lib.ia.us>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>