Bug 13537 - Focus should move to barcode after renewing via checkouts table
Koha's checkout screen automatically focuses on the barcode field each
time it is loaded. Since we've moved the checkouts table functions for
renewing and returning items to ajax methods, this page is no longer
refreshed, and thus does not refocus to the barcode after performing
those actions. This should be fixed so that renewing/returning items via
the checkouts page moves the focus to the barcode field.

Test Plan:
1) Find a patron with checkouts.
2) Renew or return a checkout, note the focus does not move to the
   barcode.
3) Apply this patch.
4) Renew or return a checkout, note the focus *does* move to the
   barcode.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>