Bug 13068: updatedatabase improvement for developer process
with this patch, the updatedatabase now checks info the atomicupdate directory. If there is some SQL here, it's executed.
Note that there is no version update, so if you run the updatedatabase script twice, you'll get an sql error probably.

This patch will be very useful for sandbox testing process.

When the RM pushes a patch with something in i'atomicupdate', he will have to copy it into updatedatabase.pl, give it a number, and remove the file in db_update

Test plan:
 * apply the patch
 * create a files with a .sql extension in the atomicupdate directory
 * run updatedatabase.pl
 * look at your database, what was in the .sql file must have been run

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>