Bug 3873: Avoid software error if the cart contains a deleted record
If a record is placed in the basket and deleted, the basket view
exploded with:
Can't use an undefined value as an ARRAY reference at
/home/koha/src/basket/basket.pl line 73.

Test plan:
1/ Add some records to the basket
2/ Delete one of them
3/ Look at your basket
It should not exploded

Note that the number of items in the basket is now wrong.

Works as expected.
Signed-off-by: Marc Veron <veron@veron.ch>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>
(cherry picked from commit 4c4c870dfee59ebe6386dbd8a7b75a55f75217f9)
Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>