BWS-PKG - Bug 11565 - decreaseLoanHighHolds needs Override
This patch allows the high holds loan length decrease to be overridden
either by a checkbox that remembers its setting during a series of
checkouts, or by a one time use override checkbox that will show
in the warning popup if a checkout is affected by high holds.

Test Plan:
1) Set up a checkout that will be affected by decreaseLoanHighHolds
2) Attempt to check out an item
3) Check the override checkbox
4) Note the checkout date is not reduced
5) Return the item
6) Start a new checkout for the patron
7) Check the "Don't decrease lean length based on holds" checkbox
8) Check out the item
9) Note you are not warned about the high holds decrease
   and that the checkout length is not reduced

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Christopher Brannon <cbrannon@cdalibrary.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 11565: (QA followup) add class to allow feature to be easily hidden

This wraps both checkboxes (in the confirmation dialog and checkout
interface) in a class of "circ-override-high-holds" so they can be
easily hidden.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 11565: QA Follo-up: Improve translatability of 'not shortened' message

Mark-up inside sentences splits them of up for translation, so it's better
to only highlight full sentences for easier translation.

Bug 11565 [QA Followup] - Hide checkout options when not in use

This follow-up reformats the checkout screen so that checkout settings
like specify due date, on-site checkout, and override high holds are
grouped in a panel which is hidden by default. Clicking a link shows all
the available settings.

Also modified in this patch: The default color for div.hint has been
made a little darker for readability.

To test, apply the patch and clear your browser cache if necessary.

- Check out to a patron who is not blocked from checking out.
- Click the "Checkout settings" link and confirm that the settings panel
  is displayed.
- Test that the settings in the panel behave as expected.
- Test with preferences on and off: decreaseLoanHighHolds,
  OnSiteCheckouts, SpecifyDueDate.

NOTE: This patch is a work in progress. It does not address the issue of
checkout setting being hidden even if they are active.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 11565 [QA Followup] - Only hide options when none are in use

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>