BWS-PKG - Bug 15736: Add ShowAllCheckins pref to control the display of checked-in item list on checking
Bug 14821 removed the items which were not checked out, but some
libraries considered it as a bug.
So let's add a new pref to control this behavior.

Test plan:
0/ Execute the updatedb entry and set ShowAllCheckins to "Do not show"
(default)
1/ Check an item in.
If the item was not checked out, it won't be listed
2/ Turn the pref to "Show"
3/ Check an item in
If the item was not checked out, it should be listed

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Bug 15736 [QA Followup] - Make add more details to the syspref description

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>